Problem 5.2-5.5

+++ src/chap05/Problem05_01.java
@@ -2,9 +2,12 @@ package chap05;
 
 /**
  * Created by yaodh on 2015/1/18.
+ *
+ * Problem 5.1
+ * How would you go about computing the parity of a very large number of 64-bit non-negative integers?
  */
 public class Problem05_01 {
-    byte parity(int x) {
+    byte parity(long x) {
         byte ans = 0;
         while (x != 0) {
             ans ^= (x & 1);
@@ -13,7 +16,7 @@ public class Problem05_01 {
         return ans;
     }
 
-    byte parity2(int x) {
+    byte parity2(long x) {
         byte ans = 0;
         while (x != 0) {
             ans ^= 1;
